+++ test/testSetup.ts
@@ -20,12 +20,19 @@ jest.mock('@expo/react-native-action-sheet');
 jest.mock('expo-font', () => {
   return {useFonts: () => []};
 });
+
 jest.mock('@react-native-async-storage/async-storage', () => mockAsyncStorage);
 jest.mock('react-native/Libraries/EventEmitter/NativeEventEmitter');
 
 jest.mock('react-native-reanimated', () =>
   require('react-native-reanimated/mock'),
 );
+
+jest.mock('react-native-device-info', () => {
+  return {
+    getDeviceTypeSync: jest.fn(),
+  };
+});
 
 // const customGlobal: any = global;
